backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & Auth  
supabase==2.0.0  # Downgraded to fix httpx client proxy issue
python-jose[cryptography]==3.3.0

# HTTP & Web Scraping
aiohttp==3.9.1
requests==2.31.0
httpx>=0.24.0,<0.25.0  # Match supabase 2.0.0 requirement
beautifulsoup4==4.12.2
feedparser==6.0.10

# NLP & ML
nltk==3.8.1  # Includes VADER sentiment analyzer
numpy==1.24.3  # Required for NLTK and data processing

# AI Services
groq==0.9.0  # For Groq AI service

# Push Notifications
exponent-server-sdk==2.0.0  # For mobile push notifications

# Database ORM
tortoise-orm==0.20.0
aerich==0.7.2  # For database migrations

# Utilities
python-dotenv==1.0.0

# Search
duckduckgo-search==3.8.5  # Compatible with httpx<0.25

# Optional Article Processing (not currently active but imported)
newspaper3k==0.2.8  # For article extraction
sumy==0.11.0  # For text summarization
lxml==4.9.3  # Required by newspaper3k
